ref: Rename the project
rename the project to `com.book.store` instead of `com.book.store.Book.Store`
