+++ src/main/java/com/book/store/Repository/OrderRepository.java
@@ -1,12 +1,11 @@
-package com.book.store.Book.Store.Repository;
+package com.book.store.Repository;
 
-import com.book.store.Book.Store.Entity.Order;
+import com.book.store.Entity.Order;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
