+++ src/main/java/com/book/store/Service/BookService.java
@@ -1,7 +1,7 @@
-package com.book.store.Book.Store.Service;
+package com.book.store.Service;
 
-import com.book.store.Book.Store.Entity.Book;
-import com.book.store.Book.Store.Repository.BookRepository;
+import com.book.store.Entity.Book;
+import com.book.store.Repository.BookRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
